% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ctrl.R
\name{add_ctrl}
\alias{add_ctrl}
\title{add_ctrl}
\usage{
add_ctrl(map, cols, rows, ..., seed = 3)
}
\arguments{
\item{map}{Experiment map to add control wells to.}

\item{cols}{Vector of plate column numbers to consider for the additional wells.}

\item{rows}{Vector of plate row numbers to consider for the additional wells.}

\item{...}{Subset of variable. The name must match a variable present in the map.}

\item{seed}{Random seed for designating remaining wells in the plate. Defaults to 3.}
}
\value{
The input map tibble, with the additional wells appended at the bottom.
}
\description{
Add wells to an experiment map over a subset of the variables in the original. Useful for adding controls, or additional independent experiments.
}
\examples{

testmath(4)
}
