% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfill.R
\name{backfill}
\alias{backfill}
\title{picklist}
\usage{
backfill(
  pl,
  backfill_wells,
  src_plate_name = "SrcPlateName",
  xfer_vol = "XferVol",
  src_well = "SrcWell",
  backfill_fluid = NA
)
}
\arguments{
\item{pl}{picklist to add back-fill transfers to}

\item{backfill_wells}{Vector of wells containing back-fill fluid in source plate}

\item{src_plate_name}{Name of source plate containing the back-fill fluid, in case multiple source plates were used to generate the picklist}

\item{xfer_vol}{= Name of the Volume column in picklist}

\item{src_well}{= Name of the Well column in the source plate}

\item{backfill_fluid}{= Name of the backfill fluid, to be assigned as the SampleName of the backfill_wells}
}
\value{
A tibble containing back-filling coordinates for the picklist provided, bringing every well to the maximum volume in the picklist.
}
\description{
Adds transfer rows of back-fill volume to a picklist generated by \code{picklist}. Returns original picklist with back-fill rows appended. All transfers in the picklist provided are assumed to take the same back-fill fluid
}
