% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_vol.R
\name{chk_vol}
\alias{chk_vol}
\title{chk_vol}
\usage{
chk_vol(pl)
}
\arguments{
\item{pl}{Picklist table. The names of the columns are assumed to be "Src_Well" and "Transfer_Volume"}
}
\value{
A tibble with the total volume transferred from each source well
}
\description{
Summarize volume transferred from each Source Well in a picklist
}
\examples{

testmath(4)

}
