% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoPickList-package.R
\docType{package}
\name{EchoPickList-package}
\alias{EchoPickList}
\alias{EchoPickList-package}
\title{EchoPickList: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
This package provides functions to design combinatorial experiments in microplates, and to export the tabels as csv picklists for use with the Echo acoustic dispenser.
}
\references{
Arroyo-Perez, EE. UdeM, 2025
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
