% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exp.R
\name{map_exp}
\alias{map_exp}
\title{map_exp}
\usage{
map_exp(cols, rows, ..., seed)
}
\arguments{
\item{cols}{Vector of plate column numbers to consider for the experiment.}

\item{rows}{Vector of plate row numbers to consider for the experiment.}

\item{...}{Variables to combine for the experiment.}

\item{seed}{Random seed for designation of wells to each condition.}
}
\value{
A tibble containing all possible combinations of the Variables in the input, and a random selection of wells assigned from the area designated by cols + rows
}
\description{
Generates a plate map for an experiment with all combinations of the defined variables, randomly assigned to the plate positions indicated in a defined area.
}
\examples{

testmath(4)
}
