% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_well.R
\name{pad_well}
\alias{pad_well}
\title{pad_well}
\usage{
pad_well(Well)
}
\arguments{
\item{Well}{A Well name string}
}
\value{
The same Well name, with zero-padding to the left of the number, to two digits.
}
\description{
Add padding zeroes to a well name. Useful for compatibility with tables generated by the \code{plater} package.
}
\examples{

testmath(4)
}
