% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picklist.R
\name{picklist}
\alias{picklist}
\title{picklist}
\usage{
picklist(
  map,
  src_plate,
  xfer_vol,
  variables,
  dest_plate = "Dest_Plate_Name",
  src_plate_name = "Src_Plate_Name",
  src_plate_type = "DMSO",
  src_well = "Well"
)
}
\arguments{
\item{map}{Experiment map to follow}

\item{src_plate}{Plate map of the source plate(s) to use for this dispense. Should be a single table.}

\item{xfer_vol}{One of the following:
Value of volume to dispense (in nL)
Vector of volumes to dispense, of length == nrow(map)
Name of a column in \code{map} corresponding to that vector}

\item{variables}{String or vector of strings with the name of the columns in \code{src_plate} to use for matching}

\item{dest_plate}{Name to assign to the Destination Plate, or the name of the column in the \code{map} containing the Destination Plate Name}

\item{src_plate_name}{Name of 'Source Plate Name' column in src_plate, or Name to be assigned}

\item{src_plate_type}{Name of 'Destination Plate type' column in src_plate, or value to be assigned. 'DMSO' and 'Water' are interpreted readily for a 384PP plate.}

\item{src_well}{Name of Source Well column in src_plate}
}
\value{
A tibble containing all possible combinations of the Variables in the input, and a random selection of wells assigned from the area designated by cols + rows
}
\description{
Generates a picklist from an experiment map and a source plate map, for use in the Beckmann Echo software suite
}
\examples{

testmath(4)
}
