% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area.R
\name{plot_area}
\alias{plot_area}
\title{plot_area}
\usage{
plot_area(cols, rows, plate_format = 384)
}
\arguments{
\item{cols}{Numeric vector of plate columns selected for the experiment.}

\item{rows}{Numeric vector of plate rows selected for the experiment.}

\item{plate_format}{Plate format used. Defaults to 384.}
}
\value{
A plot of the layout of the plate
}
\description{
GGplot2 visualizer of the plate area selected for the experiment.
}
\examples{

testmath(4)

}
