% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_vol.R
\name{round_vol}
\alias{round_vol}
\title{round_vol}
\usage{
round_vol(vol, droplet = 2.5)
}
\arguments{
\item{vol}{Vector of volumes to convert}

\item{droplet}{Size of the droplet used. Default is 2.5 nL}
}
\value{
A vector of values rounded to multiples of the droplet size
}
\description{
Rounds values to multiples of the droplet size. Default droplet is 2.5 nL
}
\examples{

round_vol(c(5, 6.5, 8, 10))

}
