% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpad_well.R
\name{unpad_well}
\alias{unpad_well}
\title{unpad_well}
\usage{
unpad_well(Well)
}
\arguments{
\item{Well}{A Well name string}
}
\value{
The same Well name, without zero-padding.
}
\description{
Remove padding zeroes from a well name. Useful for compatibility with tables generated by the \code{plater} package.
}
\examples{

testmath(4)
}
