% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfill.R
\name{backfill}
\alias{backfill}
\title{picklist}
\usage{
backfill(
  pl,
  backfill_wells,
  src_plate_name = "Src_Plate_Name",
  xfer_vol = "Transfer_Volume",
  src_well = "Src_Well"
)
}
\arguments{
\item{pl}{picklist to add back-fill transfers to}

\item{backfill_wells}{Vector of wells containing back-fill fluid in source plate}

\item{src_plate_name}{Name of source plate containing the back-fill fluid, in case multiple source plates were used to generate the picklist}
}
\value{
A tibble containing back-filling coordinates for the picklist provided, bringing every well to the maximum volume in the picklist.
}
\description{
Adds transfer rows of back-fill volume to a picklist generated by \code{picklist}. Returns original picklist with back-fill rows appended. All transfers in the picklist provided are assumed to take the same back-fill fluid
}
\examples{

testmath(4)
}
