% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoPickList-package.R
\docType{package}
\name{EchoPickList-package}
\alias{EchoPickList}
\alias{EchoPickList-package}
\title{EchoPickList: A package to design microplate experiments and to generate picklist for use with the Echo acoustic dispenser}
\description{
Contains functions to design microplate experiments with combinatorial conditions that can be added as a solution with the Echo acoustic dispenser, as well as functions to generate picklist ready to use with the Beckman Echo software.
}
\details{
This package provides functions to design combinatorial experiments in microplates, and to export the tabels as csv picklists for use with the Echo acoustic dispenser.
}
\references{
Arroyo-Perez, EE. UdeM, 2025
}
\author{
\strong{Maintainer}: Erick Arroyo-Perez \email{erickarroyo1312@protonmail.com}

}
\keyword{internal}
